requirements.txt
streamlit
gspread
pillow
requests
spotipy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
deepface
opencv-python
numpy